refactor: add fields slug, material in Product and fields is_active in all table. delete product_categories, subcategories, product_gallery
