--- app/Models/ProductGallery.php
@@ -1,21 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-
-class ProductGallery extends Model
-{
-    use HasFactory;
-
-    protected $fillable = [
-        "gallery_image",
-        "product_id"
-    ];
-
-    public function product()
-    {
-        return $this->belongsTo(Product::class);
-    }
-}
