+++ database/seeders/FakeDataSeeder.php
@@ -5,6 +5,7 @@ namespace Database\Seeders;
 use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 use Illuminate\Database\Seeder;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Str;
 
 class FakeDataSeeder extends Seeder
 {
@@ -14,32 +15,14 @@ class FakeDataSeeder extends Seeder
     public function run(): void
     {
         DB::table('categories')->insert([
-            ['name' => 'Nam'],
-            ['name' => 'Nữ'],
-            ['name' => 'Trẻ em'],
+            ['name' => 'Quần Nam'],
+            ['name' => 'Quần Nữ'],
+            ['name' => 'Quần Trẻ em'],
+            ['name' => 'Áo Nam'],
+            ['name' => 'Áo Nữ'],
+            ['name' => 'Áo Trẻ em'],
         ]);
 
-
-        $categories = DB::table("categories")->select("name", "id")->get();
-
-
-        foreach ($categories as $category) {
-            DB::table('subcategories')->insert([
-                ['name' => 'Quần ' . $category->name, "category_id" => $category->id],
-                ['name' => 'Áo ' . $category->name, "category_id" => $category->id],
-                ['name' => 'Váy ' . $category->name, "category_id" => $category->id],
-                ['name' => 'Phụ kiện ' . $category->name, "category_id" => $category->id],
-            ]);
-        }
-        $subcategories = DB::table("subcategories")->select("name", "id")->get();
-        foreach ($subcategories as $index => $subcategory) {
-            DB::table('product_categories')->insert([
-                'name' => 'product_categories ' . $index,
-                'subcategory_id' => $subcategory->id,
-                'created_at' => now(),
-                'updated_at' => now(),
-            ]);
-        }
 
         $colors = ['Red', 'Blue', 'Green', 'Yellow', 'Black', 'White'];
 
@@ -60,18 +43,22 @@ class FakeDataSeeder extends Seeder
                 'updated_at' => now(),
             ]);
         }
-        $countCategories = DB::table('product_categories')->count("id");
+        $countCategories = DB::table('categories')->count("id");
 
         for ($i = 1; $i <= 10; $i++) {
+            $sku = strtoupper(fake()->regexify('[A-Za-z0-9]{8}'));
+            $name = fake()->realText . $i;
             DB::table('products')->insert([
-                "sku" => strtoupper(fake()->regexify('[A-Za-z0-9]{8}')),
+                "slug" => Str::slug($name . '-' . $sku),
+                "material"=> "Vai",
+                "sku" => $sku,
                 'name' => 'Product ' . $i,
                 "thumbnail" => fake()->imageUrl(30, 30),
                 'short_description' => 'Short description for product ' . $i,
                 'long_description' => 'Long description for product ' . $i,
                 "regular_price" => rand(30000, 1000000),
                 "reduced_price" => rand(30000, 1000000),
-                'product_category_id' => rand(0, $countCategories),
+                'category_id' => rand(0, $countCategories),
                 'created_at' => now(),
                 'updated_at' => now(),
             ]);
